Pre prod

+++ ObaGroup.DataAccess/TokenData/OAuthTokenProperties.cs
@@ -1,83 +1,98 @@
+using System.Net;
 using System.Security.Claims;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Logging;
-using ObaGoupDataAccess.Data;
-using ObaGoupDataAccess.Repository;
 using ObaGoupDataAccess.Repository.IRepository;
-using ObaGroupModel;
+using ObaGroupUtility;
+using RestSharp;
 
 namespace ObaGoupDataAccess;
 
-public class OAuthTokenProperties
+public class OAuthTokenProperties : IOAuthTokenProperties
 {
+    private static readonly ILogger _logger = LoggerFactory.Create(builder => { builder.AddConsole(); })
+        .CreateLogger("OAuthTokenProperties");
+
+    private readonly Icryption _cryption;
     private readonly IHttpContextAccessor _httpContextAccessor;
     private readonly IUnitOfWork _unitOfWork;
-    public string accessToken;
-    
-    private static readonly ILogger _logger = LoggerFactory.Create(builder =>
-    {
-        builder.AddConsole();
-    }).CreateLogger("OAuthTokenProperties");
-    
-    public OAuthTokenProperties(IHttpContextAccessor httpContextAccessor, IUnitOfWork unitOfWork)
+
+    public OAuthTokenProperties(IHttpContextAccessor httpContextAccessor, IUnitOfWork unitOfWork, Icryption cryption)
     {
         _httpContextAccessor = httpContextAccessor;
         _unitOfWork = unitOfWork;
+        _cryption = cryption;
     }
-    
+
     public string GetAccessToken()
     {
         _logger.LogInformation("GETTING ACCESS TOKEN");
-        var context = _httpContextAccessor.HttpContext;
-        var token = context.Request.Cookies["OauthTokenAccessToken"] ?? "";
-       _logger.LogInformation("Token ="+token);
-       return token;
+        var token = new GoogleTokensUtility(_httpContextAccessor, _cryption).GetAccessToken();
+        return token;
     }
+
     public int GetExpires_in()
     {
         var context = _httpContextAccessor.HttpContext;
-        return  Int32.Parse(context.Request.Cookies["OauthExpiresIn"] ?? "0");
+        return int.Parse(context.Request.Cookies["OauthExpiresIn"] ?? "0");
     }
-    
+
     public string GetRefreshToken()
     {
         _logger.LogInformation("GETTING REFRESH TOKEN");
         var user = _httpContextAccessor.HttpContext.User;
         var userEmail = user.FindFirst(ClaimTypes.Email)?.Value;
-        _logger.LogInformation("USER EMAIL : "+userEmail);
-        UserOauthRefreshToken userOauthRefreshToken =_unitOfWork.UserOauthRefreshTokenRepository.GetFirstOrDefault(u => u.ApplicationUser.Email == userEmail);
+        _logger.LogInformation("USER EMAIL : " + userEmail);
+        var userOauthRefreshToken =
+            _unitOfWork.UserOauthRefreshTokenRepository.GetFirstOrDefault(u => u.ApplicationUser.Email == userEmail);
 
-        if (userOauthRefreshToken == null)
-        {
-            return null;
-        }
-        var refreshToken=userOauthRefreshToken.RefreshToken;
-        _logger.LogInformation("refresh token ="+ refreshToken);
+        if (userOauthRefreshToken == null) return null;
+        string? refreshToken = null;
+        if (!string.IsNullOrEmpty(userOauthRefreshToken.RefreshToken))
+            refreshToken = _cryption.Decrypt(userOauthRefreshToken.RefreshToken);
+
+
         return refreshToken;
     }
-    
+
+    public void RevokeAppAccessToCalendar(string accessToken)
+    {
+        var restClient = new RestClient();
+        var request = new RestSharp.RestRequest();
+
+        request.AddQueryParameter("token", accessToken);
+
+        restClient.BaseUrl = new Uri(Constants.Google_Revoke_Token_Endpoint);
+        var response = restClient.Post(request);
+
+        Console.WriteLine(response.Content);
+        if (response.StatusCode == HttpStatusCode.OK) RevokeToken(accessToken);
+    }
+
     public void RevokeToken(string accessToken)
     {
         _logger.LogInformation("REVOKING TOKEN");
-        var context = _httpContextAccessor.HttpContext;
-        context.Response.Cookies.Append("OauthTokenAccessToken", accessToken);
-        
-        context.Response.Cookies.Delete("OauthTokenIdToken");
-        
+
+        new GoogleTokensUtility(_httpContextAccessor, _cryption).SetAccessToken(accessToken);
+        new GoogleTokensUtility(_httpContextAccessor, _cryption).DeleteOauthTokenId();
         var user = _httpContextAccessor.HttpContext.User;
         var userEmail = user.FindFirst(ClaimTypes.Email)?.Value;
-        var userOauthRefreshTokenObj=_unitOfWork.UserOauthRefreshTokenRepository.GetFirstOrDefault(u => u.ApplicationUser.Email == userEmail);
-        _unitOfWork.UserOauthRefreshTokenRepository.Remove(userOauthRefreshTokenObj);
-        
+        var userOauthRefreshTokenObj =
+            _unitOfWork.UserOauthRefreshTokenRepository.GetFirstOrDefault(u => u.ApplicationUser.Email == userEmail);
+        if (userOauthRefreshTokenObj != null)
+        {
+            _unitOfWork.UserOauthRefreshTokenRepository.Remove(userOauthRefreshTokenObj);
+            _unitOfWork.Save();
+        }
+
         _logger.LogInformation("DONE REVOKING TOKEN");
     }
-    
-    public void overrideAccessToken(string accessTokenFromGoogle)
-    {        
-        _logger.LogInformation("OVERRIDING ACCESS TOKEN = "+accessTokenFromGoogle);
-        var context = _httpContextAccessor.HttpContext;
-        accessToken = accessTokenFromGoogle;
-        context.Response.Cookies.Append("OauthTokenAccessToken", accessToken);
+
+    public void OverrideAccessToken(string accessTokenFromGoogle)
+    {
+        _logger.LogInformation("OVERRIDING ACCESS TOKEN");
+        var accessToken = accessTokenFromGoogle;
+        new GoogleTokensUtility(_httpContextAccessor, _cryption).SetAccessToken(accessToken);
         _logger.LogInformation("DONE OVERRIDING ACCESS TOKEN");
     }
 }
